backend/requirements.txt
# GCP Cloud Functions dependencies
functions-framework==3.*

# Authentication
PyJWT==2.8.*
passlib[bcrypt]==1.7.*

# Google Cloud
google-cloud-firestore==2.14.*
google-cloud-secret-manager==2.18.*

# Utilities
python-dotenv==1.0.*

# Development dependencies (not deployed)
# pytest==7.*
# pytest-asyncio==0.21.*
# httpx==0.25.*
